Update repo struct to fix broken tests
