+++ go.mod
@@ -2,4 +2,7 @@ module github.com/blue-factory/statemachine
 
 go 1.16
 
-require github.com/sirupsen/logrus v1.8.1 // indirect
+require (
+	github.com/sirupsen/logrus v1.8.1 // indirect
+	github.com/stretchr/testify v1.7.0 // indirect
+)
